feat(rangeset): rangeset crate
