+++ rangeset/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "rangeset"
+version = "0.1.0"
+edition = "2024"
+authors = ["TLSNotary Contributors"]
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/tlsnotary/tlsn-utils"
+description = "Integer collection backed by ranges with set operation support."
+keywords = ["range", "set"]
+
+[features]
+default = []
+serde = ["dep:serde"]
+
+[dependencies]
+serde = { workspace = true, optional = true, features = ["derive"] }
+
+[dev-dependencies]
+criterion = { version = "0.5" }
+rstest = "0.12"
+itertools = "0.11.0"
+
+[[bench]]
+name = "range"
+harness = false
