+++ rangeset/README.md
@@ -1,0 +1,3 @@
+# rangeset
+
+This crate provides the `RangeSet` type which stores a collection of values as sorted disjoint ranges.
