+++ rangeset/fuzz/fuzz_targets/range_diff_range.rs
@@ -4,8 +4,8 @@ use std::ops::Range;
 
 use libfuzzer_sys::fuzz_target;
 
-use tlsn_utils_fuzz::assert_invariants;
-use utils::range::*;
+use rangeset::*;
+use rangeset_fuzz::assert_invariants;
 
 fn expected_range_difference(a: Range<u8>, b: Range<u8>) -> Vec<u8> {
     a.filter(|x| !b.contains(x)).collect::<Vec<_>>()
