+++ rangeset/fuzz/fuzz_targets/range_subset_set.rs
@@ -5,9 +5,9 @@ use std::ops::Range;
 
 use libfuzzer_sys::fuzz_target;
 
-use tlsn_utils_fuzz::SmallSet;
+use rangeset_fuzz::SmallSet;
 
-use utils::range::*;
+use rangeset::*;
 
 fuzz_target!(|r: (Range<u8>, SmallSet)| {
     let s1 = r.0;
