+++ rangeset/fuzz/fuzz_targets/set_intersection_set.rs
@@ -4,9 +4,9 @@ use std::collections::HashSet;
 
 use libfuzzer_sys::fuzz_target;
 
-use tlsn_utils_fuzz::SmallSet;
+use rangeset_fuzz::{assert_invariants, SmallSet};
 
-use utils::range::*;
+use rangeset::*;
 
 fuzz_target!(|r: (SmallSet, SmallSet)| {
     let s1: RangeSet<u8> = r.0.into();
@@ -15,5 +15,10 @@ fuzz_target!(|r: (SmallSet, SmallSet)| {
     let h1: HashSet<u8> = HashSet::from_iter(s1.iter());
     let h2: HashSet<u8> = HashSet::from_iter(s2.iter());
 
-    assert_eq!(s1.is_subset(&s2), h1.is_subset(&h2));
+    let intersection = s1.intersection(&s2);
+    let h3: HashSet<u8> = HashSet::from_iter(intersection.iter());
+
+    assert_eq!(h3, h1.intersection(&h2).copied().collect::<HashSet<_>>());
+
+    assert_invariants(intersection);
 });
