+++ rangeset/fuzz/fuzz_targets/set_union_range.rs
@@ -4,12 +4,12 @@ use std::ops::Range;
 
 use libfuzzer_sys::fuzz_target;
 
-use tlsn_utils_fuzz::{assert_invariants, SmallSet};
+use rangeset_fuzz::{assert_invariants, SmallSet};
 
-use utils::range::*;
+use rangeset::*;
 
-fn expected_union(a: Range<u8>, b: RangeSet<u8>) -> Vec<u8> {
-    let mut expected_values = a.chain(b.iter()).collect::<Vec<_>>();
+fn expected_union(a: RangeSet<u8>, b: Range<u8>) -> Vec<u8> {
+    let mut expected_values = a.iter().chain(b).collect::<Vec<_>>();
 
     expected_values.sort();
     expected_values.dedup();
@@ -21,9 +21,9 @@ fuzz_target!(|r: (Range<u8>, SmallSet)| {
     let r1 = r.0;
     let r2: RangeSet<u8> = r.1.into();
 
-    let expected_values = expected_union(r1.clone(), r2.clone());
+    let expected_values = expected_union(r2.clone(), r1.clone());
 
-    let union = r1.union(&r2);
+    let union = r2.union(&r1);
 
     let actual_values = union.iter().collect::<Vec<_>>();
