+++ spansy/src/http/types.rs
@@ -1,4 +1,4 @@
-use utils::range::{Difference, RangeSet, ToRangeSet};
+use rangeset::{Difference, RangeSet, ToRangeSet};
 
 use crate::{json::JsonValue, Span, Spanned};
