+++ utils/Cargo.toml
@@ -3,23 +3,5 @@ name = "tlsn-utils"
 version = "0.1.0"
 edition = "2021"
 
-[features]
-default = ["serde"]
-serde = ["dep:serde"]
-
 [lib]
 name = "utils"
-
-[dependencies]
-serde = { workspace = true, optional = true, features = ["derive"] }
-
-[dev-dependencies]
-criterion = { version = "0.5" }
-rstest = "0.12"
-rand = { workspace = true }
-itertools = "0.11.0"
-
-
-[[bench]]
-name = "range"
-harness = false
