+++ utils/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod filter_drain;
 pub mod id;
 pub mod iter;
-pub mod range;
 pub mod tuple;
